Moved CollectionIterator trait to new file. Remained PSR2 violations:
FILE: src/Collections/Collection.php
----------------------------------------------------------------------
FOUND 1 ERROR AND 2 WARNINGS AFFECTING 3 LINES
----------------------------------------------------------------------
  51 | WARNING | Property name "$_items" should not be prefixed with
     |         | an underscore to indicate visibility
 214 | WARNING | Property name "$_count" should not be prefixed with
     |         | an underscore to indicate visibility
 221 | ERROR   | Method name "Collection::get_Count" is not in camel
     |         | caps format
----------------------------------------------------------------------


FILE: src/Collections/CollectionIterator.php
----------------------------------------------------------------------
FOUND 0 ERRORS AND 1 WARNING AFFECTING 1 LINE
----------------------------------------------------------------------
 31 | WARNING | Property name "$_iteratorPosition" should not be
    |         | prefixed with an underscore to indicate visibility
----------------------------------------------------------------------
